+++ src/main/java/edu/ie3/util/quantities/PowerSystemUnits.java
@@ -21,7 +21,7 @@ import tech.units.indriya.unit.Units;
 /**
  * Defines the physical units used in the simulation. They can be used like this: private
  * Quantity&lt;Power&gt; activePower = Quantities.getQuantity(5.0, Units.MW); When defining
- * interfaces via Quantities.getQuanitity make sure to always use a double ( e.g. 5.0, not just 5)
+ * interfaces via Quantities.getQuantity make sure to always use a double ( e.g. 5.0, not just 5)
  * when needed or else calculations will always result in a integer. E.g.: 0.0 MW + 0.2 MW = 0.2 MW
  * but 0 MW + 0.2 MW = 0 MW
  *
@@ -85,11 +85,11 @@ public class PowerSystemUnits extends Units {
       new AlternateUnit<>(WATTHOUR_PER_METRE, "kWh/km");
 
   /** Watthour per squaremetre */
-  public static final Unit<Irradiation> WATTHOUR_PER_SQUAREMETRE =
+  public static final Unit<EnergyDensity> WATTHOUR_PER_SQUAREMETRE =
       new ProductUnit<>(WATT.multiply(HOUR).divide(SQUARE_METRE));
 
   /** Kilowatthour per squaremetre */
-  public static final Unit<Irradiation> KILOWATTHOUR_PER_SQUAREMETRE =
+  public static final Unit<EnergyDensity> KILOWATTHOUR_PER_SQUAREMETRE =
       MetricPrefix.KILO(WATTHOUR_PER_SQUAREMETRE);
 
   /* ==== Power ==== */
@@ -119,8 +119,12 @@ public class PowerSystemUnits extends Units {
   public static final Unit<Power> MEGAWATT = MetricPrefix.MEGA(WATT);
 
   /** Watt per square metre */
-  public static final Unit<PowerDensity> WATT_PER_SQUAREMETRE =
+  public static final Unit<Irradiation> WATT_PER_SQUAREMETRE =
       new ProductUnit<>(WATT.divide(SQUARE_METRE));
+
+  /** Kilowatt per square metre */
+  public static final Unit<Irradiation> KILOWATT_PER_SQUAREMETRE =
+      new ProductUnit<>(KILOWATT.divide(SQUARE_METRE));
 
   /* ==== Composed units ==== */
   public static final Unit<DimensionlessRate> PERCENT_PER_HOUR =
@@ -174,7 +178,7 @@ public class PowerSystemUnits extends Units {
   /** kW/K */
   public static final Unit<ThermalConductance> KILOWATT_PER_KELVIN = new BaseUnit<>("kW/K");
 
-  private static HashSet<String> registeredLabels = new HashSet<>();
+  private static final HashSet<String> registeredLabels = new HashSet<>();
 
   static {
     addUnit(WATTHOUR, "Wh");